work on training

+++ segmentation/src/utils.py
@@ -37,7 +37,8 @@ class CrossEntropyLoss2d(nn.Module):
             # mask = targets > 0
             targets_m = targets.clone()
             targets_m -= 1
-            loss_all = self.ce_loss(inputs, targets_m.long())
+            # loss_all = self.ce_loss(inputs, targets_m.long())
+            loss_all = self.ce_loss(inputs.float(), targets_m)
             number_of_pixels_per_class = \
                 torch.bincount(targets.flatten().type(self.dtype),
                                minlength=self.num_classes)
